requirements.txt
fastapi
uvicorn[standard]
supabase
python-dotenv
pytest
httpx

